Proposed fix for issue 6256 (stale TrackSelection reference)
Move the update of the HlsChunkSource's TrackSelection reference to be seperate from the loop that recreates HlsSampleStreams.  This way the reference is update even if the input HlsSampleStream reference was not nulled out (retain the stream when a new TrackSelection is created that is essentially a copy of the previous one.