Update to save all sample queue write indexes
Finailzed logic to update the `SampleQueue` write positions (first index) to push these into `HlsMediaChunk` when the track is initially created (from the Extractor output) and as each new chunk is queued to load (`init()` callback).

Add lots of debuging prints that can come out for the final merge.

Code is very close to a clone of `ChunkSampleStream`.

--- Backport this change from dev-v2

 Conflicts:
	demos/main/src/main/java/com/google/android/exoplayer2/demo/PlayerActivity.java
	library/hls/src/main/java/com/google/android/exoplayer2/source/hls/HlsSampleStreamWrapper.java
	library/ui/src/main/java/com/google/android/exoplayer2/ui/DebugTextViewHelper.java