Fix up back port of feature-discard-upstream-hls to trickplay mainline
Backporting the feature branch `feature-discard-upstream-hls` that is merge ready to `dev-v2` for ExoPlayer pull to work with the `t-trickplay-mainline` branch that is back at dev-v2-r2.10.4.

Note aditional changes to the PlayerActivity to support sending an intent for playback URL switch and adding a BandwidthMeter changes.

This `feature-discard-upstream-hls` is maintained as a pull request, https://github.com/google/ExoPlayer/pull/6321.  So all changes releated to fixes in the logic for it need to go in that branch and be backported.

This will be used to discard the buffered iFrames on switch back from normal.  Once this change is merged back into t-trickplay-mainline that work can begin.