Add initial seek option in demo player
Add an intent extra `start_at` to specify a number of seconds for initial playback position.  This is to allow testing like Hydra will start a playback with a saved postion.

Introduces a `MediaSourceEventCallback` that allows the `SimpleExoPlayerFacdtory` client to detect when the play URL has been iniitally loaded and playback is ready to start (or started if play when ready is true).