Final cleanup and push, now stopping iFrame playback on VTP exit
This checkin causes TrackSelection to exclude iFrame tracks from normal playback and include only iFrame tracks in trickplay selection.  This results in much more acurate entry and exit from iFrame playback.

To do this, the changes to the base `AdaptiveTrackSelection` are reverted out (as track set is already pruned).  Intead the `iFrameAwareTrackSelection` adjusts the selection to match the trick-play state.

Lastly, the pull request that discards upstream samples is included (https://github.com/google/ExoPlayer/pull/6321) in this branch.  It will be removed from the merge to t-trickplay-reverse-mainline, as it is not yet needed and still not well tested.

Lastly extra logging is removed (HlsMediaChunk, the HlsSampleStreamWrapper, etc.)