Add recovery from dead-object.
Tunneling mode on Arris (Broadcom AB4) seems to loose the audio binder that runs the audio track in tunneling mode, the AudioTrack reports a DEAD_OBJECT error (-6).

Add recovery in the SimpleExoPlayerFactory.   The recovery is intrusive, it disables the audio track then re-starts playback (with a prepare) with audio re-enabled.