Fixed bug causing VTP start delay after channel change (BZSTREAM-5239)
FIxes bug BZSTREAM-5239 by re-working seek based trickplay.  The class javadoc on the inner-class `SeekBasedTrickPlay` should describe the current algorithm.

The old algorithm drove the next seek issuance from the render events attempting to force a constant render rate.  This caused an issue when the first seek target is outside of the i-Frame playlist seekable range, with Velocix origin this is possible as they violate the HLS spec (the i-Frame variant is updated at a slower rate then the regular variants).

So the bug happens when you immediately start trickplay reverse after startup of a live stream, this causes a switch to the i-Frame playlist when Velocix may not have added i-Frame segments for the seek target, causing you to wait up to two playlist refersh cycles (over 12 seconds!).

This new algorithm has the added benefit of priortizing updating the timebar (current position) over rendering.