Fix issue with overrides in trackselection
The change to filter to i-Frame only tracks has a bug that it ignores selection overrides.   These overrides are created by the `TrackSelectionDialog` where you can either select 'auto' (removes all overrides) or one or more tracks from the group associated with a renderer (so you can limit ABR to a set of tracks).

Only the demo apps do this.   Every other override that the TiVo apps do are for non-adaptive tracks, however it is possible if we had multiple audio tracks for this to matter.