Fix the WebVTT parsing to deal with timestamp PTS wrap.
In WebVTT the sample (Cue) timestamps are allowed to ignore PTS wrap and monotonically increase in time.   Current code will not see this problem unless playing a live stream that has run for more then 29 hours (such that PTS wraps but timestamps do not).

This is described in a single obtuse paragraph in the
Pantos spec [section 3.5](https://tools.ietf.org/html/draft-pantos-hls-rfc8216bis-04#section-3.5)

Quoted below:

```
 When synchronizing WebVTT with PES timestamps, clients SHOULD account
   for cases where the 33-bit PES timestamps have wrapped and the WebVTT
   cue times have not.  When the PES timestamp wraps, the WebVTT segment
   SHOULD have a X-TIMESTAMP-MAP header that maps the current WebVTT
   time to the new (low valued) PES timestamp.
```

The change is basically a no-op if the timestamps never wrap.