Update to honor CaptioningManager Locale
The Android Accessability settings UI allows the user to specify default for
captions (on/off) as well as the preferred language for captions. The default for this is to use the system language.

This code change watches for settings changes to the captioning style,
enablement, and preferred language and update the track selection when they change.

Code like this can "fix" https://jira.tivo.com/browse/PARTDEFECT-2791, assuming the user
are ok with setting their preferred captioning language as a solution.