Recover from audiotrack init failure (v557 error)
Add recovery logic for Android AudioTrack initialization errors, the logic retries with or
without tunneling enabled.   The logic has been tested for HDMI hot plug on Xaiomi-MI,
Arris, Amino, Pyxis (SEI), FireTV (old FireOS 5.2.7.3)

This "fixes" (or works arround more accurately:

1. Switching Audio Track on live TV causes v557 - PARTDEFECT-2072
2. Non-secure Tunneling not supported for AC3 tracks (Amino, Arris, others)
3. Audio track write failures on HDMI removal (this occurs on Amino / Arris, not  Pyxis (SEI500), MI or FireTV)

It is possibe to use the `DefaultExoPlayerErrorHandler` without the entire `SimpleExoPlayerFactory`.  All that
is required is to create an instance of the `DefaultExoPlayerErrorHandler`, add it as an analytics listener and
add the `PlaybackExceptionRecovery` implementers of your choice.  See the `createPlayerErrorHandler()` method in
`SimpleExoPlayerFactory` for an example of how this is done.

 Conflicts:
   SimpleExoPlayerFactory.java - took ours, more recent version of audio track recovery