Fix so heeader parse does not break test cases
Apparently comments and other garbage are legal after the WEBVTT, also an empty vtt file must report as the file is invalid.

This is a cherry-pick of the one commit (fixing the \t in the header) that we wanted but Google will not
accept.  There is a bug (https://jira.tivo.com/browse/PARTDEFECT-2618) once this is fixed, we can remove
this commit.