CEA-708 Decoder fixes for issue #1807.
In this commit -

Handling the sequence number discontinuity in caption channel packet header.

Handle the multiple Service Blocks within a single caption channel packet. An outer while loop is added to function processCurrentPacket() to achieve that.

The processCurrentPacket returns if the packet length does not match with the currentIndex. That assumption is wrong.

As per spec the the packet can end on reception of next cc_type = 0x3.

Added a logic to handle multi byte commands. These commands can be within the service block or spread across service blocks.

Processing of service block needs to wait until all the command bytes arrive. An ArrayList is used to accumulate the bytes belonging to same service block.

In this change the current flow is not altered, i.e. the parsing of service block is not started until the complete caption channel packet arrives.

These changes are verified through Sarnoff Test streams. Many tests are still failing after this change. Subsequent changes are planned to address that.

This is a cherry-pick of one commit from a pending pull request:
  https://github.com/google/ExoPlayer/pull/8356
And the commit was:
  https://github.com/google/ExoPlayer/pull/8356/commits/93829f94057f1c2c1309d880df3ef67dca10e1eb

The other two commits in the pull request are already included, cherry-picked from Google's 1.15.1

The conflicts resolved are:
  * use previousSequenceNumber over lastSequenceNumber
  * take Google's spacing choices
  * Google's code for the anchorId to Cue.ANCHOR_TYPE to match 2.15.1