CEA-708 Decoder fixes for issue #1807.
Fix requires changes to the merged pull request https://github.com/google/ExoPlayer/pull/8415
that are still part of the unmerged pull request https://github.com/google/ExoPlayer/pull/8356

In this commit -

Handling the sequence number discontinuity in caption channel packet header.

Handle the multiple Service Blocks within a single caption channel packet. An outer while loop is added to function processCurrentPacket() to achieve that.

The processCurrentPacket returns if the packet length does not match with the currentIndex. That assumption is wrong.

As per spec the the packet can end on reception of next cc_type = 0x3.

Added a logic to handle multi byte commands. These commands can be within the service block or spread across service blocks.

Processing of service block needs to wait until all the command bytes arrive. An ArrayList is used to accumulate the bytes belonging to same service block.

In this change the current flow is not altered, i.e. the parsing of service block is not started until the complete caption channel packet arrives.

These changes are verified through Sarnoff Test streams. Many tests are still failing after this change. Subsequent changes are planned to address that.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# Author:    sneelavara <sneelavara@gmail.com>
# Date:      Wed Dec 16 13:03:37 2020 -0800
#
# On branch t-fix-cea708Merge
# Your branch and 'tivo-pvt/t-fix-cea708Merge' have diverged,
# and have 1 and 2 different commits each, respectively.
#   (use "git pull" to merge the remote branch into yours)
#
# Changes to be committed:
#	modified:   library/core/src/main/java/com/google/android/exoplayer2/text/cea/Cea708Decoder.java
#
# Untracked files:
#	CCURStream_3000kbps720p29_97fps-1_1661286386_init.cmfv?ccur_keyrot_t=1666096234
#	CCURStream_3000kbps720p29_97fps-1_T1666098015585589~D6006000.cmfv
#	MERGE_COMMIT_MESSAGE
#	VelocixVTP.mp4
#	X
#	test_url
#	tivo-docs/DASH Encoding Guidelines.md
#	tivo-docs/Encoding HLS Renditions.pdf
#	tivo-docs/HANDLING_2.15.1_MERGE_CONFLICTS.pdf
#	tivo-docs/Logging Updates For Audio Position.md
#	tivo-docs/Logging Updates For Audio Position.pdf
#	tivo-docs/TiVoExoPlayerReleaseMapping.md
#	tivo-docs/TiVoExoPlayerReleaseMapping.pdf
#	tivo-docs/Understanding ExoPlayer Position Tracking.md
#	tivo-docs/Understanding ExoPlayer Position Tracking.pdf
#	widevine.cmfv
#