Add playback end reason and tests cases
Added determining the playback end reason (`PlaybackMetrics.EndReason` enum).  This uses the final event to determine if we reached the end of content or ended because of an error or channel change.

Added test cases and/or boasted coverage to certify the following PlaybackMetrics:

* `getAvgVideoBitrate()` -- playing bitrate excluding trickplay
* `getAvgNetworkBandwidth()` -- bandwidth measured for media loads (audio, video or captions)
* `getTotalTrickPlayTime()` -- total time in ms playing in trick-play mode since start of session
* `getTrickPlayCount()` -- count of the trickplay operations
* `getProfileShiftCount()` -- number of ABR (profile shifts), excluding shifts for trick-play
* `getTimeInVideoFormat()` -- time playing in each variant (excluding time in trick-play

Other methods in `PlaybackMetrics` are not yet unit tested.