Update API to allow integration with Client-Core
The client-core code creates it's PlaybackMetrics object in a static method
called rom Haxe that then uses an inner class (Runnable queued to run on UI
thread) to update the metrics.  This requires the API to provide split methods:

1. `PlaybackMetrics createOrReturnCurrent()`
2. `updateFromCurrentStats(PlaybackMetrics metrics)`

Where 1 can be called form Haxe thread and 2 can (only) be called from the
Android UI thread.

With this update, the code has been locally integrated with client/core hydra and
is ready to release with 2.11.6-2.1.