Dual Mode VTP Phase 0 (#130)
* Dual Mode Phase VTP Phase 0 w/o reverse optimizations

Includes the alternate playlist generation with test cases. Tests pass
and lint is clean (one small version issue)

With this version reverse playback performs on par with release 2.11.6-2.1
Note reverse requires a delicate balance of:

1. Dense i-Frame playlist to find a frame to render
2. Enough buffered to allow the renderer to start producing frames

Each seek clears the sample queues and triggers a load, this is built into the architecture

This is essentially these two commits:

* b2311443d3 2021-03-08 | Fixes bug where odd playlist length causes null segment (t-dual-mode-vtp-support) [Steve Mayhew]
* da0145e381 2021-01-27 | WIP: proof of concept using parser factory for dualmode [Steve Mayhew]

With rebase to release (2.11.6-2.2)
So this checkin replaces the unpushed branch `t-dual-mode-vtp-support`

* Buffered now reported correctly for live playlists

Fixed an issue where the generate live i-Frame only playlist was marked with end tag, this
 caused buffered duration to report bogus numbers (no surprise, mixing live and non-live variants).

Also updated test case with change to reverse playback track selection.

* Removes un-needed override from Renderer

Now that scan mode and scrub mode already curate the playlist to affect a
desired frame density there is no need to limit the reads to the upstream source.
This code was required when many of the i-Frames would be discarded if frame rate is exceeded.

* Trickplay and IFrameAwareAdaptiveTrackSelection are thread safe...

The IFrameAwareAdaptiveTrackSelection methods are called from the player thread, so must be careful
to only call thread safe methods in TrickPlayController (which is updated from the UI thread).

Refactored creation of IFrameAwareAdaptiveTrackSelection.Factory to the
TrickPlayControlFactory, this allows hiding it from clients (ExoPlayerPlayer and the like)

Bump version to V3.x as this is a pretty major change to the API (player / trick-play creation)

* Format label is printed in prefrence over id

This allows Dual Mode created variants to be easy to identify, they are only
differnt by label from the source variant

* Fix bug where TrackSelection did not change in FF mode with long availableDuration

Once again, adding a test case finds a bug ;-).   The `IFrameAwareAdaptiveTrackSelection`
Must override the logic that blocks track selection changes based on buffering.  The code had chosen the wrong value to get buffered duration.

Also, refactored the playlist subset value to a constant and added it to the label so it is clear what subset is chosen