Makes metrics objects type hierarchy match use case (#136)
* Makes metrics objects type hierarchy match use case

With this change the `TrickPlayMetrics` are not polluted with vacuous values that are germane only to `PlaybackMetrics`.  The base class, `AbstractBasePlaybackMetrics` contains the shared values.

This allows us to expand metrics in a more rational fashion going forward.

Note: Hydra imports import com.tivo.exoplayer.library.metrics.PlaybackMetrics.EndReason explicitly this
causes a compile error if it is defined in a base class, so `TrickPlayMetrics` borrows this enum, while less than 
perfect it allows this to work with current client core glue code.  Trick Play either ends with an error or normal, so 
the other EndReasons are not used.