Dockerized build that installs Android SDK on based JDK11
Using the based JDK-11 linux docker (`docker.tivo.com/openjdk-docker:11`)
Install the Android SDK tools and android platform tools in a fashion that
allows gradle build of an Android APK

The `Dockerfile` is suitable for building ExoPlayer, but could be re-used for
any Android build.

NOTE: Gradle scripts and plugins depend on Java system properties to determine
where cache files, gradle installs, etc.  In particular the user and their
home directory.  The the Dockerfile includes a "build" user with the UID/GID
of the user on the host that runs the build.  This allows java's `user.home`
property to come out correct (not "?") when the container is run with -u by
the jenkins docker command (which passes the builder UID/GID to allow writing
the bind mounted WORKDIR)