no infinite stall if GAP > maxBufferMs
This is a back-port of the commit in the pull request to Google in dev-v2.
That commit is (4ebf25a6)[https://github.com/TiVo/ExoPlayer/commit/87e80dc4643f3bd4eba6fc0974fdcb2e4ebf25a6]
in TiVo's public ExoPlayer fork.

this fix is the same code execpt for the change to `ExoPlayerImplInternal.java` as that is not
required to support version before 2.14.x of ExoPlayer

This fixes it by never counting completed GAP tag segments in the buffered position calculation