Fix issue where a trun atom could be associated with the wrong track
Note that this removes a workaround for malformed content, in which the
track_ID is set incorrectly. It's unclear there was sufficient reason to
implement that workaround, and so it's preferable to remove it, rather
than implementing the concept of unrecognized tracks, which would be
needed to keep it and to also fix this issue.

Issue: #9056
PiperOrigin-RevId: 379506261

RESOLVED CONFLICTs
 - removed function `getTrackBundle()`, logic in that function was inlined by change that added `FLAG_SIDELOADED`
 - revert changes to RELEASENOTES.md, will pickup in our own commit