handle PlaylistStuckException with rational retry
Add an error recovery handler to the library to handle the `PlaylistStuckException`.  This error occurs when the origin stops transcoding new segments for longer then the 2.5 * the target duration (about 14 seconds).

The recovery logic retries but keeps in buffering state and does not replay previously played content.

Also refactored and added methods to the `PlaybackExceptionRecovery` interface to enable this kind of handling, this includes:

1. Add to the API for PlayerErrorHandlerListener to report error and progress recovering
2. remove dependencies from error handlers on components other than `PlayerErrorRecoverable`
3. added test cases