Fixes BZSTREAM-7913, NPE when video is off
This issue, BZSTREAM-7913  "App crashes when press back from screen saver. exoplayer2.Format.label' on
a null object reference" is caused because the playing time in formats reports a null format for playback
while video render is disabled (this is as designed, I missed it).

The fix includes updating our EventLogger logic to move it to a subclass, with this we can revert all changes
to base ExoPlayer EventLogger accept for opening a few methods as protected.

Other related changes:
1. Add API for mobile (or anyone) to replace the EventLogger subclass used by SimpleExoPlayerFactory
1. Convert construction of factory to builder pattern
1. Updated triage document with minor format changes after EventLogger subclass.