Add method for MediaSourceLiveCycle to cleanup
Add a method to allow `MediaSourceLiveCycle` to cleanup references.

Note the use case for the `SimpleExoPlayerFactory` preserves everything across
channel (MediaSource, playUrl) changes.  Only when you call `releasePlayer()` is everything cleaned up.  Removing listeners in this is not so needed, as long as the only reference to the player itself is the one being nulled in the `releasePlayer()` method.

NOTE WELL - this is part of the reason the `getTrackSelector()` is deprecated, the object this method returns can and *will* hold on to reference to the `ExoPlayerImplInternal`  wired to the ExoPlayer object via a `Handler` and it's `Looper`