Fixes WSIPCL-11330 - FFWD stuck at SOCU start
A fix an and a workaround for [WSIPCL-11330](https://jira.tivo.com/browse/WSIPCL-11330).

Change includes a workaround for Vecima issue,
`EXT-X-PROGRAM-DATE-TIME` mismatch across variants.  The workaround truncates the time to second boundary.
This caused the first mentioned issue "The FFWD trickplay was stuck at the beginning"

The workaround is not ideal, but it fixes the issue.

Secondly, the `Handler` is not always run down properly for rewind (this is because the render event
is delivered on the player thread).  The fix for this is in our code, `TrickPlayController`.
This is causing the issue "sometime saw the [forward] trickplay getting jumped to the beginning  of cache"

Also deprecated API's from ExoPlayer used in `TrickPlayController` are replaced with non-deprecated equivalents.