More reliable renders for `ScrubTrickPlay`
1. Update lastPosition from the discontinuity event (so it includes adjustments)
2. Set nearest seek based on direction of move.
3. Renderer ignores decode only in scrub mode

For point 2, we Initially use CLOSEST_SYNC, subsequentmoves
backwards/forwards use `PREVIOUS_SYNC`/`NEXT_SYNC` respectively.

For point 3,  If the origin mismatches the timestamp in an iFrame (the PTS in
the IDR is earlier then the current timeline time) then `SampleStream` marks it
the sample as decodeonly (this is done because non-IDR samples still need to be
decoded to produce the picture, but not displayed)

During trickplay we want to display every IDR we get so ignore this flag.