Prevent referencing a dead player in error recovery
This change prevents referencing a player that has been released in error recovery.

Specifically, replace the deprecated `resetAndRetryPlayback()` with `retryPlayback()`,
which is the only supported way to retry in ExoPlayer 2.12 and beyound.

Also, remove the related method in `MediaSourceLiveCycle` and deference the `player` object
 in this class.  The long term plan is to deprecate `MediaSourceLiveCycle` as ExoPlayer 2.12
 and beyond are increasing hiding the `MediaSource` in favor of using `MediaItem` APIs

Thanks @spvishwas122 and @adroitandroid (Prateek) for the ground work for this fix,
 which they expect will fix issue https://jira.tivo.com/browse/BZQSA-4341