Removes listener on recovery abort or player release.
It is possible for clients to call `SimpleExoPlayer.releasePlayer()` which error recovery is active.  In this case recovery should abort and cleanup any state.

TODO: will add some test cases around this, note with lastest (2.12) ExoPlayer it is possible to be notified with `onPlayerError()` even after the player is released, with `ExoPlaybackException.TIMEOUT_OPERATION_RELEASE`  so clients should be prepared for this.

TODO:  also should remove our Player.EventListeners after `player.release()` returns