Uses correct index for playlist URL
The TrackSelection API returns two indexes, the index of the selected
track (`getSelectedIndex()`) and the index in the `TrackGroup`.
The `HlsChunkSource` and  `HlsSampleStreamWrapper` should only work with the later.

Actually the `getSelectedIndex()`, current selected track index is really only useful
on the API to determine if there is a valid selection or not.
The index is really internal to `TrackSelection`