Does not attempt recover from release timeout.
Our 2.12 release pulls forward a change that sets a timer on `Player.release()`.
If that timer expires `onPlayerError()` is called.

Attempting any recovery from this error is pointless, the player is already
completely rundown.   The timeout occurs when the video renderer code
calls MediaCodec.stop() and the call does not return in 500ms

The issues related to this is:
 https://github.com/google/ExoPlayer/issues/4352
    //  and commit: