Log but do not report release timeout
This change fixes BZSTREAM-8839 and friends (crash in onStop() in ExoPlayerPlayer).

What is happening is the backport of the Google fix for Issue: #4352 (https://github.com/google/ExoPlayer/issues/4352) reports a playerError on release.  While this is interesting to log, there is nothing that the code can do to recover from it.

The code in ExoPlayerPlayer would recover / report the error then call `Player.release()`, of course it was already in stop so this error report would cause unexpected reentry into `ExoPlayerPlaeyer.stop()` resulted in a the NPE.