First draft of Adaptive Dual Mode design
This design spec outlines the API changes and internal design of the new feature to
allow adatpive track selection for trick-play.

In a nut shell the change permits the video track selection to remain part of the
resulting TrackGroup then dynamically filtered by the `AdaptiveTrackSelection`

This was part of the original design negotiated with Oliver see
[this comment](https://github.com/google/ExoPlayer/issues/474#issuecomment-514307811)
but was not implemented quite that way because there was no support for
flushing trickplay tracks.