Fix for WSIPCL-12725, regression for buffering state in tunneling mode
Stall detection in tunneling mode was originally implemented in commit
3ec50ba297396e1880f4df7fbd5cd9e0ffe3a7d8 by adding a lastOutputTimeUs variable
and checking the difference between lastInputTimeUs and lastOutputTimeUs.
Meanwhile, commit 29c43b5e241f6e8cd774e2aa85e24f9472c5af95 refactored
MediaCodecVideoRenderer to move lastInputTimeUs to the parent class. Merge
commit 8e227a1eabbef166ecbee7caa8b743d1e54db361 notes the conflict, and chooses
to keep the TiVo fix, including the addition of lastOutputTimeUs, as well as
adding lastInputTimeUs back in. However, it misses a line in onQueueInputBuffer
where lastInputTimeUs was actually set, making the stall detection fail. This
fix instead follows the refactor and replaces uses of lastInputTimeUs with calls
to getLargestQueuedPresentationTimeUs, fixing the regression. This is not an
ideal fix, since it still deviates from upstream, which will require a more
careful approach than timestamp tracking.