Non tunneled Arris reverse renders frames
This fix is a work around for the issue where BRCM CODECs require multiple
frames queued in order to decode and produce the first frame queued.

Simply queues the same iFrame multiple times.  Small rendering artifacts occur which i may be possible to eliminate (maybe don't render the duplicates?)

The bug is https://jira.xperi.com/browse/PARTDEFECT-11896

So, if that is fixed this work-around can be removed.