WSIPCL-12356: Almost all channels some time the trick play FF restarts from the beginning
Method HlsChunkSource.getNextChunk allows the track switch to happen when the new target
playlist is not yet available (is being refreshed). Thus, when fetching the playlist isn't
too fast, the first segment may come from the old playlist snapshot. That is okay as
its position is determined by media time. The next segment, however, will be picked up on
mediaSequence-plus-one basis. If the fresh playlist is fully switched at this point,
getNextChunk will return incorrect segment or will jump beyond live window. This change is
to minimize that risk