Remove tunneling mode override of hasOutputReady()
With the 1 second value for requiredBufferedTimeUs in AdaptiveLoadControl.shouldContinueLoading() this should not be needed. Originally I was using a very small value to try and ensure only 1 frame would be downloaded even if the iframe spacing didn't conform to the encoding spec. At that time it would sometimes stall when it encountered iframe with a duration of less than 333ms which is occurring in Vecima 4K streams. With the 1 second requiredBufferedTimeUs there's a chance it could show 2 close together frames during a single seek but that should not be a big deal and we also have a bug against Vecima to fix it.