Workaround to handle 5-channel AC3 audio.
Many Android devices don't support passthrough of 5-channel AC3, so we're assuming 5-channel AC3 means 6-channel even when lfeon bit is not set.
When the content is really 5 channels or  6 but with the lfeon bit not set, on all platforms tested so far playback succeeds when we specify 6-channel and on many platforms,
Also when I create 5-channel AC3 content on the problem platforms it will only succeed if we specify 6 channels.
We are also working with Vecima to fix the Dolby 5.1 channels to correctly set the lfeon bit.
We are also still working with SEI and Broadcom to see if the platform can be made to handle 5-channel AC3 when requested to do so.
Also while the AC3 spec may allow 5 channel audio, in reality, this is probably only seen in the wild due to an encoder config mistake.