Add handling for iframe transport segments with filler data.
When the H264Reader downloads an iframe segment with filler data at the end, the filler data is considered
a NAL (type 12) and becomes the last NALU.  This breaks the code that will commit the sample if the last NAL is an IDR.

Fixed this by assuming we're done when we hit filler data. The filler data should only be at the end is of no interest.
Since the fIller data is just that, fix is simple, add the filler data (basically all FF bytes) to the sample
and don’t end it as a stand alone NAL unit in the parsing loop