Validates playlist times to MS only
The `isUpdateValid()` is a bit overzealous to check to the nearest micro-second.

Vecima PROGRAM-DATE-TIME is only to millisecond boundary (of course Pantos
is ambiguous on this one, showing example with milliseconds but format string
example does not include.  ISO8601 is specified, but for this milliseconds are
optional.

Duration is even worse, it should be "floating point", wonderful.

This change lightens up to allow time to the nearest millisecond to be ok

One might argue that Vecima should publish times that match in resolution.