Supports dual-mode playlist updates
Update to not regenerate the curated playlist on update, instead update the curated
playlist in place based on changes to the source playlists.

The previous handling for this genreated updated curated playlists that did not reflect the
changes to the source playlist correctly.  This lead to issues like BLWE which caused
the bug:
  WSIPCL-12356 [DualModeTP] On all channels sometimes the trick play FF restarts from the beginning