Fix for paused playback on exit from VTP (wsipcl-13498)
ExoPlayerPlayer checks the setPlayWhenReady flag to determine the pause state.
A race condition condition where a pending trickFrameRendered() is delivered
after `scrubStop()` is called will result in paused playback on exit from trick play.

player.setPlayWhenReady(false) in the rendered callback is not required as this
must already be the state, removing this fixes the race.

Fix is for bug https://jira.xperi.com/browse/WSIPCL-13498