iFrame ABR track selection adapts on target frame rate
Use the `FrameRateAnalyzer`, exposed through the `TrickPlayControlInternal` to
determine the best `Format` for iFrame only trick-play and select it. The `FrameRateAnalyzer`
measures the average frame rate for iFrame tracks when the playlist is loaded.

The algorithm finds the track with the closest frame rate to the requested frame rate for
the current trickplay speed.  Frame rate must be measured from

If no playlists have been loaded, defaults to start with original source, once this is loaded
and measured the next segments may switch to load from the nearer match.