IFrameAwareAdaptiveTrackSelection honors overrides correctly
Currently, the IFrameAwareAdaptiveTrackSelection customization does not support setMaxVideoSize() and setMaxVideoBitrate() APIs,
or any other constraints that result in overrides that limit the tracks in the `TrackGroup` to a subset.  The logic will incorrectly
include the iFrame only tracks in the set during normal playback (fails to filter the iFrame only tracks)

As part of a player stress management policy a version of track selection restriction support requires limiting the tracks.  In
addition, the current track selection will elimiate 4k tracks where screens do not have 4k support, this will trigger this
bug that leaks iFrame only tracks.