Updates library-tivo-ui to Google 2.15.1
The primary changes in 2.15.1 affecting library-tivo-ui are:

1. `PlaybackException` was lifted as a super-class of `ExoPlaybackException`
2. `TrackSelection` migrated to `ExoTrackSelection`
3. Changes to `AnalyticsCollector` to introduce a `Handler` for delivering multiple events to
   `AnalyticsListener` interface broke the test cases for our analitics package.

For the first change, simply use `PlaybackException` everywhere until it is possible and required to
downcast to `ExoPlaybackException` (eg the `AudioTrackInitPlayerErrorHandler`)

The second change is to simply use `ExoTrackSelection` in places where we know the track
selection is dynamic (e.g. `StressBasedTrackSelector`).
Long term we need to follow the path that seems to be laid in place by Google commit:
[*Split mutations method out of TrackSelection*](https://github.com/google/ExoPlayer/commit/ec437350543fa43763b7ba74e0b43ec37a1ee224)

The last change, for AnalyticsCollector will require more work, for now the test cases are marked @Ignore

Updates library-tivo-ui to Google 2.15.1

The primary changes in 2.15.1 affecting library-tivo-ui are:

1. `PlaybackException` was lifted as a super-class of `ExoPlaybackException`
2. `TrackSelection` migrated to `ExoTrackSelection`
3. Changes to `AnalyticsCollector` to introduce a `Handler` for delivering multiple events to
   `AnalyticsListener` interface broke the test cases for our analitics package.
4. Update to remove dependency on `@MediaCodecRenderer.MediaCodecOperationMode`, this symbol
   is removed in 2.15.1.  Replace with a simple boolean and a TODO to match the new renderer
   implementation in `SimpleExoPlayerFactory` API

For the first change, simply use `PlaybackException` everywhere until it is possible and required to
downcast to `ExoPlaybackException` (eg the `AudioTrackInitPlayerErrorHandler`)

The second change is to simply use `ExoTrackSelection` in places where we know the track
selection is dynamic (e.g. `StressBasedTrackSelector`).
Long term we need to follow the path that seems to be laid in place by Google commit:
[*Split mutations method out of TrackSelection*](https://github.com/google/ExoPlayer/commit/ec437350543fa43763b7ba74e0b43ec37a1ee224)

The third change, for AnalyticsCollector will require more work, for now the test cases are marked @Ignore

For the fourth change, the new API in the ExoPlayer Renderer for async operation must be plumbed in
to the call SimpleExoPlayerFactory.