First checkin of demo-tenofoot for 2.15.1
With this checkin the tenfoot demo builds correctly with 2.15.1 and the
updates to the TiVo libraries.

A basic smoke test of VTP works correctly and the issue with track selection is
fixed and working properly.

More extensive testing of this demo player is the next steps, while we prepare
client-core / Hydra changes.