Add Live Offset feature of ExoPlayer 2.15.1 to tenfoot demo.
We use the MediaItem.Builder returned from the factoriesCreated() callback
to set the Live Target Offset to 30s.  As is, this change uses the default
values for minOffsetMs, maxOffsetMs, minPlaybackSpeed and maxPlaybackSpeed
which gives us a chance to try out the "dynamic" adjusting of the offset.

To disable the dynamic behavior, we would set both minPlaybackSpeed
and maxPlaybackSpeed to 1.0f.

Please refer to https://exoplayer.dev/live-streaming.html for more info.