Commits sample on end of single non-IDR iframe stream
We've encountered iframe track encodings with single iframe segments with a non-IDR iframe.
This also looks to be a valid random access points due to it being proceeded by an SEI with frame_cnt=0.
Because we are only applying this code at end of stream when the iframe is the only thing in the stream,
this change does not also check for the SEI with frame_cnt=0. I can't think of any reason
releasing the data from a single iframe stream would be problematic regardless of IDR iframe or not.
Also limiting the commit on eos non-IDR iframe behavior to stream that contain ADUs.