Discards incompatible frame rate on speed change
The logic to discard buffered segments in the method `IFrameAwareAdaptiveTrackSelection.evaluateQueueSize()`
was not discarding on both speed increase and decrease, this would cause too fast
frame rates, the change fixes these issues:

1. [WSIPCL-15564](https://jira.xperi.com/browse/WSIPCL-15564) - FF frame rates are too fast (especially FFx3)
2. [WSIPCL-15602](https://jira.xperi.com/browse/WSIPCL-15602) - FFx3 freezes for about 10 seconds at times during ...

This is not unique to  2.15.1, but perhaps prior to this version the extra frames were simply
being discareded rather then rendered.

The changes to 2.15.1 for `LoadControl` to add the live offset caused the `AdaptiveLoadControl.shouldStartPlayback()`
to be overly agressive in buffering to start forward trickplay, the change for this fixes:

1. [WSIPCL-15563](https://jira.xperi.com/browse/WSIPCL-15563) - FFx1 has delay of 3-4 seconds when initiated

Both of these will have better fixes in the next phase of dual mode VTP that will
greatly reduce the buffering and alter the way frame-rate is managed.